The baseline JIT compiler design involves tracing an EBB when its usage frequency exceeds a predetermined threshold. We then utilize a code generator to convert the instruction sequence into C code. The C code is subsequently compiled using mir, and the resulting target machine code is stored in the code cache for future utilization.

The primary objective of introducing the baseline JIT compiler is to enhance the execution speed of RISC-V instructions with minimal modification. This implementation requires three additional components: a code generator, mir compiler, and code cache, while maintaining the original design of the interpreter. Furthermore, this baseline JIT compiler serves as the foundational target for future improvements.

In addition, we have developed a Python script that effectively traces code templates and automatically generates JIT code templates. This approach eliminates the need for manually writing duplicated codey.

* Core i7-8700

| Metric    |     master      |  baseline JIT   |Speedup |
|-----------|-----------------|-----------------|--------|
| CoreMark  |1352.843 (iter/s)|1932.496 (iter/s)| +42.8% |
| Dhrystone |1282 DMIPS       |2495.11 DMIPS    | +94.6% |

* eMAG

| Metric    |     master      |  baseline JIT   |Speedup |
|-----------|-----------------|----------------|--------|
| CoreMark  | 438.073 (iter/s)| 644.426 (iter/s)| +47.1% |
| Dhrystone | 354.60 DMIPS    | 867.00 DMIPS    |  +144% |

Known issue: this implementation does not correctly work on Apple M1 MacOS.